requirements.txt
mediapipe.solutions
numpy
opencv-python
insightface
onnxruntime
